+++ packages/blocks/src/blocks/core-list.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { cn, getStylesFromAttributes } from '@snapwp/core';
 import type { CoreList as CoreListType, CoreListProps } from '@snapwp/types';