+++ packages/blocks/src/blocks/core-paragraph.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { getStylesFromAttributes } from '@snapwp/core';
 import { Parse } from '@snapwp/next';
 import type {