+++ packages/blocks/src/blocks/core-pattern.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import type {
 	CorePattern as CorePatternType,
 	CorePatternProps,