+++ packages/blocks/src/blocks/core-separator.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { getClassNamesFromString } from '@snapwp/core';
 import type {
 	CoreSeparator as CoreSeparatorType,