+++ packages/next/src/components/font.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 /**
  * A wrapper for loading fonts dynamically using next/font.
  *