+++ packages/next/src/config/with-snap-wp.ts
@@ -26,6 +26,7 @@ const modifyWebpackConfig = ( snapWPConfigPath: string ) => {
 	 *
 	 * @return The modified webpack configuration.
 	 */
+	// eslint-disable-next-line @typescript-eslint/no-explicit-any -- Using `any` type as the parameter type is `any` in Next.js.
 	return ( config: any ) => {
 		const configPath = `
 			import __snapWPConfig from '${ snapWPConfigPath }';
@@ -111,9 +112,14 @@ const withSnapWP = async ( nextConfig?: NextConfig ): Promise< NextConfig > => {
 	setConfig();
 	const homeUrl = new URL( getConfig().homeUrl );
 
+	const userImages = nextConfig?.images ?? {};
+	const userRemotePatterns = userImages.remotePatterns ?? [];
+
 	return {
 		...nextConfig,
 		images: {
+			// User image config is appended before default config. Otherwise default remote patterns will be overridden
+			...userImages,
 			remotePatterns: [
 				{
 					protocol: 'http',
@@ -123,6 +129,7 @@ const withSnapWP = async ( nextConfig?: NextConfig ): Promise< NextConfig > => {
 					protocol: 'https',
 					hostname: homeUrl.hostname,
 				},
+				...userRemotePatterns,
 			],
 		},
 		webpack: modifyWebpackConfig( snapWPConfigPath ),