+++ packages/next/src/root-layout/global-head.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import Fonts from '@/components/font';
 import { type GlobalHeadProps } from '@snapwp/core';
 
@@ -22,7 +21,7 @@ export function GlobalHead( {
 		<>
 			{ globalStylesheet && (
 				<style
-					id={ 'stylesheet' }
+					id="stylesheet"
 					dangerouslySetInnerHTML={ { __html: globalStylesheet } }
 				></style>
 			) }
@@ -33,7 +32,7 @@ export function GlobalHead( {
 
 			{ customCss && (
 				<style
-					id={ 'customCSS' }
+					id="customCSS"
 					dangerouslySetInnerHTML={ { __html: customCss } }
 				></style>
 			) }