+++ packages/next/CHANGELOG.md
@@ -1,5 +1,34 @@
 # Changelog
 
+## 0.2.1
+
+### Patch Changes
+
+-   [#112](https://github.com/rtCamp/snapwp/pull/112) [`ade29f3`](https://github.com/rtCamp/snapwp/commit/ade29f32cc3215c25e447c785eb864f081ffce4d) Thanks [@Swanand01](https://github.com/Swanand01)! - feat: Display error fallback when there are no editor blocks
+
+-   [#113](https://github.com/rtCamp/snapwp/pull/113) [`d7b87dc`](https://github.com/rtCamp/snapwp/commit/d7b87dc2b2dc0e0221fa074cdd951e39ed44a0ab) Thanks [@ayushnirwal](https://github.com/ayushnirwal)! - fix: remote patterns in users config being ignored
+
+-   [#108](https://github.com/rtCamp/snapwp/pull/108) [`c1ee556`](https://github.com/rtCamp/snapwp/commit/c1ee5564e1046411dac0ccb5ace99c59fe337959) Thanks [@Pathan-Amaankhan](https://github.com/Pathan-Amaankhan)! - chore: Enforce "import/default" eslint ruleset.
+
+-   Updated dependencies [[`c1ee556`](https://github.com/rtCamp/snapwp/commit/c1ee5564e1046411dac0ccb5ace99c59fe337959)]:
+    -   @snapwp/core@0.2.1
+
+## 0.2.0
+
+### Minor Changes
+
+-   [#103](https://github.com/rtCamp/snapwp/pull/103) [`579ba54`](https://github.com/rtCamp/snapwp/commit/579ba54e72d78b8b5fae89c8ddcd6bbdc0487f91) Thanks [@ashutoshgautams](https://github.com/ashutoshgautams)! - refactor!: Change `location` to `groupLocation` for script placement enum type
+
+### Patch Changes
+
+-   [#87](https://github.com/rtCamp/snapwp/pull/87) [`71bd814`](https://github.com/rtCamp/snapwp/commit/71bd81452935736170e81dabe8fea48b6d2b8085) Thanks [@BhumikP](https://github.com/BhumikP)! - dev: Implement noPropertyAccessFromIndexSignature rule in tsconfig base
+
+-   [#95](https://github.com/rtCamp/snapwp/pull/95) [`58d3d47`](https://github.com/rtCamp/snapwp/commit/58d3d4755962456347295dd7d876d02486a8a455) Thanks [@justlevine](https://github.com/justlevine)! - dev: Add compatibility for `ScriptModuleDependency.importType` type change to `ScriptModuleImportTypeEnum`
+
+-   Updated dependencies [[`dcecabf`](https://github.com/rtCamp/snapwp/commit/dcecabfa9df535727e988d2db59bb0a6aa5d2a73), [`71bd814`](https://github.com/rtCamp/snapwp/commit/71bd81452935736170e81dabe8fea48b6d2b8085), [`579ba54`](https://github.com/rtCamp/snapwp/commit/579ba54e72d78b8b5fae89c8ddcd6bbdc0487f91)]:
+    -   @snapwp/core@0.2.0
+    -   @snapwp/query@0.2.0
+
 ## 0.1.0
 
 ### Minor Changes