+++ packages/types/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @snapwp/types
 
+## 0.1.2
+
+### Patch Changes
+
+-   [#108](https://github.com/rtCamp/snapwp/pull/108) [`c1ee556`](https://github.com/rtCamp/snapwp/commit/c1ee5564e1046411dac0ccb5ace99c59fe337959) Thanks [@Pathan-Amaankhan](https://github.com/Pathan-Amaankhan)! - chore: Enforce "import/default" eslint ruleset.
+
+## 0.1.1
+
+### Patch Changes
+
+-   [#87](https://github.com/rtCamp/snapwp/pull/87) [`71bd814`](https://github.com/rtCamp/snapwp/commit/71bd81452935736170e81dabe8fea48b6d2b8085) Thanks [@BhumikP](https://github.com/BhumikP)! - dev: Implement noPropertyAccessFromIndexSignature rule in tsconfig base
+
 ## 0.1.0
 
 ### Minor Changes